Feat loyalty program (#44)
* feat(Backend): Optimistic lock implemented

* fix(Frontend): Show components for users points and categories control at admin system settings page

* fix(Backend): Save period in appointment instead of duration. Create models for users categories and point settings. Impl of repo, service, and controller related to new classes.

* fix(Frontend): Show points, user category for the user, and add registration reason in the registration form

* fix(Frontend): Show modification buttons on the service page just for the service owner

* feat(Frontend): Create points settings form at system page and disabled admin response to registration request if the email is not confirmed

* fix(Frontend): Fixed formating

* fix(Backend): Fix tutor service controller and modification in points settings and user category settings controllers

* fix(Backend): Points and registration reason in user dtos

* ref(Backend): Changed names of exception classes by naming convention

* fix/ref(Backend): Refactored model classes, add new validations, adding periods to tutor instead of tutor service

* feat(Backend): Points settings and user category repositories

* fix(Backend): Invalidate jwt when admin confirms delete request or user change pass

* fix(Backend): Create CompleatedAppointment and add points to users

* feat(Backend): Points settings and users category services

+++ Backend/FishingBooker/FishingBooker/src/main/java/com/isa/FishingBooker/mapper/calendar/YearCalendarMapper.java
@@ -26,6 +26,6 @@ public class YearCalendarMapper<E extends Appointment>{
 
 	@SuppressWarnings("deprecation")
 	private double calculateReservedDays(Date date, List<E> entities) {
-		return entities.stream().filter(item-> item.getStart().getMonth()==date.getMonth()).mapToDouble(Appointment::getDuration).sum();
+		return entities.stream().filter(item-> item.getStart().getMonth()==date.getMonth()).mapToInt(Appointment::getDuration).sum();
 	}
 }
